+++ NonUniqueValueException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace steevanb\PhpTypedArray;
+
+class NonUniqueValueException extends \Exception
+{
+}