+++ changelog.md
@@ -1,4 +1,11 @@
-### 1.0.1 - 2018-08-05
+### [1.1.0](../../compare/1.0.1...1.1.0) - 2018-08-05
+
+- Add `AbstractTypedArray::merge()`
+- Add `AbstractTypedArray::setUniqueValues()` and `AbstractTypedArray::isUniqueValues()` 
+- Add `AbstractTypedArray::setExceptionOnNonUniqueValue()` and `AbstractTypedArray::isExceptionOnNonUniqueValue()`
+- Add `AbstractTypedArray::__construct()` `$uniqueValues` and `$exceptionOnNonUniqueValue` parameters
+
+### [1.0.1](../../compare/1.0.0...1.0.1) - 2018-08-05
 
 - Change array type hint to iterable in `AbstractTypedArray::__construct()`